requirements.txt
geopandas
landsatxplore
rasterio
pandas
matplotlib
numpy
seaborn
fastapi
uvicorn
pydantic
tarfile
pillow
seaborn